Weapon, skill, fight dto added

+++ Services/CharacterSevices/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using DOTNET_RPG.Data;
 using DOTNET_RPG.Dtos.Character;
 using DOTNET_RPG.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace DOTNET_RPG.Services.CharacterSevices
@@ -15,21 +17,26 @@ namespace DOTNET_RPG.Services.CharacterSevices
       
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        public CharacterService(IMapper mapper, DataContext context)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
             _mapper = mapper;
+            _httpContextAccessor = httpContextAccessor;
+
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
-            
+            character.user = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync(); 
 
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.Characters.Where(c => c.user.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
@@ -39,11 +46,18 @@ namespace DOTNET_RPG.Services.CharacterSevices
             try
             {
 
-                Character character = await _context.Characters.FirstAsync(c => c.ID == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                Character character = await _context.Characters
+                 .FirstOrDefaultAsync(c => c.ID == id && c.user.Id == GetUserId());
+                 if(character != null){
+                      _context.Characters.Remove(character);
+                      await _context.SaveChangesAsync();
+                      serviceResponse.Data = (_context.Characters.Where(c => c.user.Id == GetUserId())
+                       .Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                 }    
+                 else{
+                     serviceResponse.Success = false; 
+                     serviceResponse.Message = "Character not found";
+                 } 
             }
             catch (Exception ex)
             {
@@ -55,10 +69,10 @@ namespace DOTNET_RPG.Services.CharacterSevices
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacter = await _context.Characters.Where( c=> c.user.Id == userId).ToListAsync();
+            List<Character> dbCharacter = await _context.Characters.Where( c=> c.user.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacter.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -66,7 +80,11 @@ namespace DOTNET_RPG.Services.CharacterSevices
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character selected = await _context.Characters.FirstOrDefaultAsync(c => c.ID == id);
+            
+            Character selected = await _context.Characters
+            .Include(c=>c.Weapon)
+            .Include(c=>c.CharacterSkills).ThenInclude(cs=> cs.Skill)
+            .FirstOrDefaultAsync(c => c.ID == id && c.user.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(selected);
             return serviceResponse;
         }
@@ -76,19 +94,26 @@ namespace DOTNET_RPG.Services.CharacterSevices
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
+                Character character = await _context.Characters.Include(c=> c.user).FirstOrDefaultAsync(c => c.ID == updateCharacter.ID);
+                if(character.user.Id == GetUserId())
+                {
+                    character.Name = updateCharacter.Name;
+                    character.Intelligence = updateCharacter.Intelligence;
+                    character.Defense = updateCharacter.Defense;
+                    character.Class = updateCharacter.Class;
+                    character.HitPoints = updateCharacter.HitPoints;
+                    character.Strength = updateCharacter.Strength;
 
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.ID == updateCharacter.ID);
-                character.Name = updateCharacter.Name;
-                character.Intelligence = updateCharacter.Intelligence;
-                character.Defense = updateCharacter.Defense;
-                character.Class = updateCharacter.Class;
-                character.HitPoints = updateCharacter.HitPoints;
-                character.Strength = updateCharacter.Strength;
+                    _context.Characters.Update(character);
+                    await _context.SaveChangesAsync();
 
-                _context.Characters.Update(character);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else{
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
+                
             }
             catch (Exception ex)
             {
